% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements_dir.R
\docType{data}
\name{dir_setter_panel}
\alias{dir_setter_panel}
\title{dirPanel}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
dir_setter_panel
}
\description{
dirPanel
}
\keyword{datasets}
